+++ lib/sequel-rails/setup.rb
@@ -4,10 +4,7 @@ module Rails
   module Sequel
 
     def self.setup(environment)
-      conf = configuration.environment_for(environment.to_s).reverse_merge({
-        logger: configuration.logger
-      })
-      ::Sequel.connect conf
+      ::Sequel.connect configuration.environment_for environment.to_s
     end
 
   end