+++ spec/spec_helper.rb
@@ -21,9 +21,13 @@ end
 
 # Ensure db exists and clean state
 begin
-  db = Sequel::Model.db
-  all_tables = (db.tables-[:schema_info, :schema_migrations]).map{|t| %["#{t}"]}.join ','
-  db.run "TRUNCATE TABLE #{all_tables} RESTART IDENTITY" unless all_tables.empty?
+  require "sequel-rails/storage"
+  Rails::Sequel::Storage.new(:test).drop
+  Rails::Sequel::Storage.new(:test).create
+
+  require 'sequel/extensions/migration'
+  load "#{Rails.root}/db/schema.rb"
+  Sequel::Migration.descendants.first.apply Sequel::Model.db, :up
 rescue Sequel::DatabaseConnectionError => e
   puts "Database connection error: #{e.message}"
   puts "Ensure test db exists before running specs."